refactor(core): drop the byte alias, its fields are plain int now
byte был объявлен как char (structs.h), и это ловушка: fmt печатает char
символом, а не числом. В medit (#3837) поле "F) NumDamDice" из-за этого
выводилось пустым -- значение 3 уходило в терминал управляющим символом ETX.
В диффе такое не видно вообще, ловится только на живом выводе, и наступить
на это может кто угодно, кто переведёт очередной sprintf на fmt.

Поэтому убран сам тип: alias byte удалён из structs.h, все объявления
переведены на int.

Затронуты:

* поля мобов -- damnodice, damsizedice, extra_attack, like_work,
  last_direction;
* поля комнаты -- light, glight, gdark; поля дескриптора -- bad_pws,
  idle_tics;
* таблицы спас-бросков в pc_classes.cpp и возвращаемые значения
  GetSavingThrows/GetExtendSavingThrows;
* attach_type у триггера, cleanup_type и тип записи в OLC, find_invis в
  sub_write, пол и минимальная позиция в биндингах скриптов;
* приведения на загрузке мира (yaml/sqlite/legacy) и в admin API: было
  clamp<byte>/static_cast<byte>, стало clamp<int>, лишние приведения убраны.

EGender был объявлен как "enum class EGender : char" с комментарием, что это
подогнано под byte. Подгонять больше не подо что, а char как основа enum'а --
та же ловушка при to_underlying, так что тоже int.

Заодно убраны ставшие ненужными static_cast<int> в medit и yaml-выгрузке
(часть комментариев "// byte -> int" там врала и раньше: mem_queue.total,
GoldNoDs и GoldSiDs уже давно int) и static_cast<ubyte> в animate_dead, где
целое присваивалось int-полю через 8 бит.

Что осталось намеренно: sbyte (int8_t) и ubyte (uint8_t). Они форматируются
как числа и fmt, и printf, а ubyte-массивы SplMem/SplKnw на каждого персонажа
-- это по паре килобайт, которые в int превратятся в восемь.

Проверено вживую: на тестовом мире снят вывод "смотреть", stat room, stat mob,
"опознать" и обход меню medit -- на master и на этой ветке. Различий нет.

Сборка чистая, 712 тестов зелёные.

+++ src/engine/entities/char_data.h
@@ -153,20 +153,20 @@ struct char_special_data {
 
 // Specials used by NPCs, not PCs
 struct mob_special_data {
-	byte last_direction;    // The last direction the monster went
+	int last_direction;    // The last direction the monster went
 	int attack_type;        // The Attack Type Bitvector for NPC's
 	EPosition default_pos;    // Default position for NPC
 	mob_ai::MemoryRecord *memory;    // List of attackers to remember
-	byte damnodice;        // The number of damage dice's
-	byte damsizedice;    // The size of the damage dice's
+	int damnodice;        // The number of damage dice's
+	int damsizedice;    // The size of the damage dice's
 	std::array<int, kMaxDest> dest;
 	int dest_dir;
 	int dest_pos;
 	int dest_count;
 	int activity;
 	BitsetFlags<ENpcFlag> npc_flags;
-	byte extra_attack;
-	byte like_work;
+	int extra_attack;
+	int like_work;
 	int MaxFactor;
 	int GoldNoDs;
 	int GoldSiDs;
